requirements.txt
streamlit>=1.34.0
python-dotenv>=1.0.1
pymongo>=4.7.0
google-generativeai>=0.8.0
